fix decode functions would accept uint16_t instead of uint8_t

+++ gameboy/include/gameboy/cpu/cpu.h
@@ -80,8 +80,8 @@ private:
     void on_if_write(const address16&, uint8_t data) noexcept;
     [[nodiscard]] uint8_t on_if_read(const address16&) const noexcept;
 
-    [[nodiscard]] uint8_t decode(uint16_t inst, standard_instruction_set_t);
-    [[nodiscard]] uint8_t decode(uint16_t inst, extended_instruction_set_t);
+    [[nodiscard]] uint8_t decode(uint8_t inst, standard_instruction_set_t);
+    [[nodiscard]] uint8_t decode(uint8_t inst, extended_instruction_set_t);
 
     void set_flag(flag flag) noexcept;
     void reset_flag(flag flag) noexcept;
